+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `VerticalStepper`.
+- Feature: Add `WarningIcon` and `WaitingIcon`.
+- Feature: Update `CheckedIcon` to accept `color` and `title` props.
+
 ## [1.14.0] - 2019-02-12
 
 Features: